losts of changes to sync and more

+++ android/src/main/java/com/la/runners/activity/adapter/RunListCursorAdapter.java
@@ -19,7 +19,7 @@ public class RunListCursorAdapter extends CursorAdapter {
 
 	@Override
 	public void bindView(View view, Context context, Cursor cursor) {
-		Float distance = Model.Run.distance(cursor);
+		Long distance = Model.Run.distance(cursor);
 		TextView distanceTextView = (TextView)view.findViewById(R.id.runListActivity_distance);
 		distanceTextView.setText("" + distance);
